Add offset to list of Statements in a Block
This probably will need to be upgraded to an entire PosState or similar
but start working toward having source context information attached to
parsed objects in the concrete language tree structure. For now we'll
limit it to an annotation outside Statement in the Block as
[(Offset,Statement)].

Reporting an error to the nearest statement (which to a first
approximation is one Expression per line) should be a good start.