Experiment using ParseErrorBundle
Since **megaparsec** has support for custom error types, and already has
the machinery for "pretty" printing, experiment with using it to also
print the error messages from later stages in the compiler.

Having plumbed source locations through Statements, build all the
adaptors to take a FailureReason and use it to construct a FancyError.
Wrap all this up in a ParseErrorBundle so that later `errorBundlePretty`
can be used to print it out.

This isn't working. The offsets aren't being used as we'd expect, and
the code to turn offset into line and column isn't being invoked as we'd
hoped.

This is a failed experiment, unfortunately. We're going to have to do
this ourselves.