Colourize unexpected and expected tokens
When encountering a compilation error in the parsing phase, colour the
unexpected token (if there is one) and the suggestion of what tokens are
expected.

Since this is all internal to Technique.Failure, rearrange things to
carry the ErrorItem Char on the FailureReason message and then convert
directly to Doc TechniqueToken.

Takes advantage of the Pretty instance for NonEmpty. Thanks, whoever
added that.