Change primitive action type to Program
Rename Context (in Technique.Evaluator) to Unique, mostly to stop using
the name Context which comes from Core.Program. It represents a unique
invocation of a given Executable, so sure.

Change the action type in the Primitive constructor for Function to be a
`Program None Value` rather than `IO Value`.